refactor: internalize utilities for backwards compatibility
Internalize utility functions to reduce the risk of introducing
backward compatibility issues in the public-facing API when
refactoring the codebase.
