+++ src/gbif_registrar/register.py
@@ -1,19 +1,14 @@
 """Functions for registering datasets with GBIF."""
 import os.path
 import tempfile
-import json
-import requests
 import pandas as pd
-from gbif_registrar.config import (
-    GBIF_API,
-    INSTALLATION,
-    ORGANIZATION,
-    PASSWORD,
-    USER_NAME,
+from gbif_registrar._utilities import (
+    _get_local_dataset_endpoint,
+    _expected_cols,
+    _get_local_dataset_group_id,
+    _get_gbif_dataset_uuid,
+    _read_registrations_file,
 )
-from gbif_registrar.utilities import read_registrations_file
-from gbif_registrar.utilities import get_local_dataset_endpoint
-from gbif_registrar.utilities import expected_cols
 
 
 def initialize_registrations_file(file_path):
@@ -62,7 +57,7 @@ def initialize_registrations_file(file_path):
     if os.path.exists(file_path):
         pass
     else:
-        data = pd.DataFrame(columns=expected_cols())
+        data = pd.DataFrame(columns=_expected_cols())
         data.to_csv(file_path, index=False, mode="x")
 
 
@@ -86,7 +81,7 @@ def register_dataset(local_dataset_id, registrations_file):
     --------
     >>> register_dataset("edi.929.2", "registrations.csv")
     """
-    registrations = read_registrations_file(registrations_file)
+    registrations = _read_registrations_file(registrations_file)
     # If the local_dataset_id already exists in the registrations file, then
     # there's no work to be done, so return the registrations file as-is.
     if local_dataset_id in set(registrations["local_dataset_id"]):
@@ -113,7 +108,7 @@ def register_dataset(local_dataset_id, registrations_file):
         with tempfile.NamedTemporaryFile(suffix=".csv", delete=True) as temp_file:
             new_record.to_csv(temp_file.name, index=False, mode="w")
             complete_registration_records(temp_file.name)
-            new_record = read_registrations_file(temp_file.name)
+            new_record = _read_registrations_file(temp_file.name)
         registrations = pd.concat([registrations, new_record], ignore_index=True)
         registrations.to_csv(registrations_file, index=False, mode="w")
     return None
@@ -140,7 +135,7 @@ def complete_registration_records(registrations_file):
     --------
     >>> complete_registration_records("registrations.csv")
     """
-    registrations = read_registrations_file(registrations_file)
+    registrations = _read_registrations_file(registrations_file)
     # Get all rows where the registrations dataframe columns
     # local_dataset_group_id, local_dataset_endpoint, gbif_dataset_uuid,
     # is_synchronized contain empty values. These are the rows
@@ -158,33 +153,33 @@ def complete_registration_records(registrations_file):
     # Iterate through the rows of the record dataframe.
     for index, row in record.iterrows():
         # If the local_dataset_group_id column is empty, then call the
-        # get_local_dataset_group_id function to get the local_dataset_group_id
+        # _get_local_dataset_group_id function to get the local_dataset_group_id
         # value and insert it into the local_dataset_group_id column of the
         # registrations dataframe.
         if pd.isnull(row["local_dataset_group_id"]):
-            local_dataset_group_id = get_local_dataset_group_id(
+            local_dataset_group_id = _get_local_dataset_group_id(
                 local_dataset_id=row["local_dataset_id"]
             )
             # Add the local_dataset_group_id value to the
             # local_dataset_group_id column of the registrations dataframe.
             registrations.loc[index, "local_dataset_group_id"] = local_dataset_group_id
         # If the local_dataset_endpoint column is empty, then call the
-        # get_local_dataset_endpoint function to get the local_dataset_endpoint
+        # _get_local_dataset_endpoint function to get the local_dataset_endpoint
         # value and insert it into the local_dataset_endpoint column of the
         # registrations dataframe.
         if pd.isnull(row["local_dataset_endpoint"]):
-            local_dataset_endpoint = get_local_dataset_endpoint(
+            local_dataset_endpoint = _get_local_dataset_endpoint(
                 local_dataset_id=row["local_dataset_id"]
             )
             # Add the local_dataset_endpoint value to the
             # local_dataset_endpoint column of the registrations dataframe.
             registrations.loc[index, "local_dataset_endpoint"] = local_dataset_endpoint
         # If the gbif_dataset_uuid column is empty, then call the
-        # get_gbif_dataset_uuid function to get the gbif_dataset_uuid value and
+        # _get_gbif_dataset_uuid function to get the gbif_dataset_uuid value and
         # insert it into the gbif_dataset_uuid column of the registrations
         # dataframe.
         if pd.isnull(row["gbif_dataset_uuid"]):
-            gbif_dataset_uuid = get_gbif_dataset_uuid(
+            gbif_dataset_uuid = _get_gbif_dataset_uuid(
                 local_dataset_group_id=registrations.loc[
                     index, "local_dataset_endpoint"
                 ],
@@ -195,112 +190,3 @@ def complete_registration_records(registrations_file):
             registrations.loc[index, "gbif_dataset_uuid"] = gbif_dataset_uuid
     registrations.to_csv(registrations_file, index=False, mode="w")
     return None
-
-
-def get_local_dataset_group_id(local_dataset_id):
-    """Return the local_dataset_group_id value.
-
-    Parameters
-    ----------
-    local_dataset_id : str
-        The dataset identifier in the EDI repository. Has the format:
-        {scope}.{identifier}.{revision}.
-
-    Returns
-    -------
-    str
-        The local_dataset_group_id value.
-    """
-    # The local_dataset_group_id value is derived by dropping the last period
-    # and everything after it from the local_dataset_id value.
-    local_dataset_group_id = local_dataset_id.rsplit(".", 1)[0]
-    return local_dataset_group_id
-
-
-def get_gbif_dataset_uuid(local_dataset_group_id, rgstrs):
-    """Return the gbif_dataset_uuid value.
-
-    Parameters
-    ----------
-    local_dataset_group_id : str
-        The dataset group identifier in the EDI repository. Has the format:
-        {scope}.{identifier}.
-    rgstrs : pandas dataframe
-        The registrations file as a dataframe.
-
-
-    Returns
-    -------
-    str
-        The gbif_dataset_uuid value. This is the UUID assigned by GBIF to the
-        local dataset group identifier. A new value will be returned if a
-        gbif_dataset_uuid value doesn't already exist for a
-        local_dataset_group_id.
-
-    Notes
-    -----
-    The local_dataset_group_id and gbif_dataset_uuid values have a one-to-one
-    relationship because this allows a dataset series (i.e. multiple versions
-    of a dataset) to be registered with GBIF as a single dataset and displayed
-    from a single URL endpoint on the GBIF system.
-
-    Examples
-    --------
-    >>> registrations = read_registrations_file("registrations.csv")
-    >>> get_gbif_dataset_uuid("edi.929", registrations)
-    """
-    # Look in the rgstrs dataframe to see if there is a matching
-    # local_data_set_group_id value, and it has a non-empty gbif_dataset_uuid
-    # value. If so, get the gbif_dataset_uuid value.
-    has_group_id = local_dataset_group_id in rgstrs["local_dataset_group_id"].values
-    has_uuid = (
-        rgstrs.loc[
-            rgstrs["local_dataset_group_id"] == local_dataset_group_id,
-            "gbif_dataset_uuid",
-        ]
-        .notnull()
-        .iloc[0]
-    )
-    if has_group_id and has_uuid:
-        gbif_dataset_uuid = rgstrs.loc[
-            rgstrs["local_dataset_group_id"] == local_dataset_group_id,
-            "gbif_dataset_uuid",
-        ].iloc[0]
-    # If there is no matching local_dataset_group_id value, or if there is a
-    # matching local_dataset_group_id value, but it has an empty
-    # gbif_dataset_uuid value, then call the register_dataset function to
-    # register the dataset with GBIF and get the gbif_dataset_uuid value.
-    else:
-        gbif_dataset_uuid = request_gbif_dataset_uuid()
-    return gbif_dataset_uuid
-
-
-def request_gbif_dataset_uuid():
-    """Request a GBIF dataset UUID value from GBIF.
-
-    Returns
-    -------
-    str
-        The GBIF dataset UUID value. This is the UUID assigned by GBIF to the
-        local dataset group.
-    """
-    title = "Placeholder title, to be written over by EML metadata from EDI"
-    data = {
-        "installationKey": INSTALLATION,
-        "publishingOrganizationKey": ORGANIZATION,
-        "type": "SAMPLING_EVENT",
-        "title": title,
-    }
-    headers = {"Content-Type": "application/json"}
-    resp = requests.post(
-        url=GBIF_API,
-        data=json.dumps(data),
-        auth=(USER_NAME, PASSWORD),
-        headers=headers,
-        timeout=60,
-    )
-    if resp.status_code != 201:
-        print("HTTP request failed with status code: " + str(resp.status_code))
-        print(resp.reason)
-        return None
-    return resp.json()
