+++ tests/conftest.py
@@ -1,46 +1,7 @@
 """Configure the test suite."""
 
 import pytest
-from gbif_registrar.utilities import read_registrations_file
-
-
-@pytest.fixture(name="rgstrs")
-def rgstrs_fixture():
-    """Read the test registrations file into DataFrame fixture."""
-    return read_registrations_file("tests/registrations.csv")
-
-
-@pytest.fixture(name="local_dataset_id")
-def local_dataset_id_fixture():
-    """Create a local_dataset_id fixture for tests."""
-    # This local_dataset_id corresponds to a data package archived on EDI and
-    # formatted according to the DwCA-Event core.
-    return "edi.929.2"
-
-
-@pytest.fixture
-def local_dataset_group_id():
-    """Create a local_dataset_group_id fixture for tests."""
-    # This local_dataset_group_id corresponds to a data package archived on
-    # EDI and formatted according to the DwCA-Event core. This value is
-    # derived from the local_dataset_id fixture and should be the same as the
-    # local_dataset_id with the version number removed.
-    return "edi.929"
-
-
-@pytest.fixture(name="local_dataset_endpoint")
-def local_dataset_endpoint_fixture():
-    """Create a local_dataset_endpoint fixture for tests."""
-    # This local_dataset_endpoint corresponds to a data package archived on
-    # EDI and formatted according to the DwCA-Event core.
-    return "https://pasta.lternet.edu/package/download/eml/edi/929/2"
-
-
-@pytest.fixture(name="gbif_dataset_uuid")
-def gbif_dataset_uuid_fixture():
-    """Create a gbif_dataset_uuid fixture for tests."""
-    # This gbif_dataset_uuid is an example UUID returned by the GBIF API.
-    return "4e70c80e-cf22-49a5-8bf7-280994500324"
+from gbif_registrar._utilities import _read_registrations_file
 
 
 @pytest.fixture(name="eml")
@@ -61,6 +22,13 @@ def eml_fixture():
     return xml_content
 
 
+@pytest.fixture(name="gbif_dataset_uuid")
+def gbif_dataset_uuid_fixture():
+    """Create a gbif_dataset_uuid fixture for tests."""
+    # This gbif_dataset_uuid is an example UUID returned by the GBIF API.
+    return "4e70c80e-cf22-49a5-8bf7-280994500324"
+
+
 @pytest.fixture(name="gbif_metadata")
 def gbif_metadata_fixture():
     """Create a dict of GBIF metadata for testing."""
@@ -71,3 +39,35 @@ def gbif_metadata_fixture():
         ],
     }
     return metadata
+
+
+@pytest.fixture(name="local_dataset_endpoint")
+def local_dataset_endpoint_fixture():
+    """Create a local_dataset_endpoint fixture for tests."""
+    # This local_dataset_endpoint corresponds to a data package archived on
+    # EDI and formatted according to the DwCA-Event core.
+    return "https://pasta.lternet.edu/package/download/eml/edi/929/2"
+
+
+@pytest.fixture
+def local_dataset_group_id():
+    """Create a local_dataset_group_id fixture for tests."""
+    # This local_dataset_group_id corresponds to a data package archived on
+    # EDI and formatted according to the DwCA-Event core. This value is
+    # derived from the local_dataset_id fixture and should be the same as the
+    # local_dataset_id with the version number removed.
+    return "edi.929"
+
+
+@pytest.fixture(name="local_dataset_id")
+def local_dataset_id_fixture():
+    """Create a local_dataset_id fixture for tests."""
+    # This local_dataset_id corresponds to a data package archived on EDI and
+    # formatted according to the DwCA-Event core.
+    return "edi.929.2"
+
+
+@pytest.fixture(name="rgstrs")
+def rgstrs_fixture():
+    """Read the test registrations file into DataFrame fixture."""
+    return _read_registrations_file("tests/registrations.csv")
